requirements.txt
accelerate==1.8.1
deepspeed==0.17.2
peft==0.17.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
transformers==4.57.1

numpy==1.26.4
pandas==2.2.3
scikit-learn==1.6.1
tqdm==4.67.1

librosa==0.11.0
soundfile==0.13.1

av==15.1.0
decord==0.6.0
opencv-python==4.11.0.86
Pillow==11.3.0
matplotlib==3.10.3

PyYAML==6.0.2
safetensors==0.5.3
toolz==1.1.0
qwen-omni-utils==0.0.8
